move css files into styles folder (they're not components!)
