+++ components/WeekDisplay.js
@@ -1,5 +1,5 @@
 import React, {  useState, useEffect } from "react";
-import "./WeekDisplay.css"
+import "./styles/WeekDisplay.css"
 import WbSunnyIcon from '@material-ui/icons/WbSunny';
 import CloudIcon from '@material-ui/icons/Cloud';
 import Grain from '@material-ui/icons/Grain';
